refactor: extract action handler to separate module (Phase 13.1)
- Create handler/action.rs with:
  - ActionResult enum for action execution results
  - EntrySnapshot struct for entry data
  - ActionContext struct (extracted from Config)
  - handle_action function for KeyAction execution
  - get_target_directory and get_filename_str helpers
- Add unit tests for the new module
- Update main.rs to use the extracted module
- Export new types from handler/mod.rs

This refactoring enables testing of action execution logic
separately from the main event loop.
