+++ src/main.rs
@@ -19,13 +19,17 @@ use crossterm::{
 use ratatui::prelude::*;
 use ratatui::widgets::{Block, Borders, Paragraph};
 
-use fileview::action::{file as file_ops, Clipboard, ClipboardContent};
-use fileview::core::{AppState, InputPurpose, PendingAction, ViewMode};
+use fileview::action::file as file_ops;
+use fileview::core::{AppState, ViewMode};
 use fileview::handler::{
-    key::{create_delete_targets, handle_key_event, update_input_buffer, KeyAction},
+    action::{
+        get_filename_str, get_target_directory, handle_action, ActionContext, ActionResult,
+        EntrySnapshot,
+    },
+    key::{handle_key_event, update_input_buffer},
     mouse::{handle_mouse_event, ClickDetector, MouseAction, PathBuffer},
 };
-use fileview::integrate::{exit_code, Callback, OutputFormat, PickResult};
+use fileview::integrate::{exit_code, Callback, OutputFormat};
 use fileview::render::{
     is_binary_file, is_image_file, is_text_file, render_directory_info, render_hex_preview,
     render_image_preview, render_input_popup, render_status_bar, render_text_preview, render_tree,
@@ -200,37 +204,6 @@ fn run() -> anyhow::Result<i32> {
     result
 }
 
-/// Snapshot of entry data for use after dropping borrow
-#[derive(Clone)]
-struct EntrySnapshot {
-    path: PathBuf,
-    name: String,
-    is_dir: bool,
-    depth: usize,
-}
-
-/// Get the target directory for file operations.
-/// If the focused path is a directory, use it directly.
-/// Otherwise, use its parent directory or fall back to root.
-fn get_target_directory(focused: Option<&PathBuf>, root: &Path) -> PathBuf {
-    focused
-        .and_then(|p| {
-            if p.is_dir() {
-                Some(p.clone())
-            } else {
-                p.parent().map(|pp| pp.to_path_buf())
-            }
-        })
-        .unwrap_or_else(|| root.to_path_buf())
-}
-
-/// Get the filename from a path as a string for display purposes.
-fn get_filename_str(path: Option<&PathBuf>) -> String {
-    path.and_then(|p| p.file_name())
-        .map(|n| n.to_string_lossy().to_string())
-        .unwrap_or_default()
-}
-
 /// Handle file drop operation - copy files to target directory.
 /// Returns the number of files successfully processed.
 fn handle_file_drop(
@@ -278,6 +251,12 @@ fn run_app(
     let mut navigator = TreeNavigator::new(&config.root, state.show_hidden)?;
     let mut click_detector = ClickDetector::new();
     let mut path_buffer = PathBuffer::new();
+
+    // Create action context from config
+    let action_context = ActionContext {
+        callback: config.callback.clone(),
+        output_format: config.output_format,
+    };
 
     // Preview cache
     let mut text_preview: Option<TextPreview> = None;
@@ -520,7 +499,7 @@ fn run_app(
                         &mut navigator,
                         &focused_path,
                         &snapshots,
-                        &config,
+                        &action_context,
                         &mut text_preview,
                     )? {
                         ActionResult::Continue => {}
@@ -596,385 +575,4 @@ fn run_app(
             return Ok(exit_code::SUCCESS);
         }
     }
-}
-
-enum ActionResult {
-    Continue,
-    Quit(i32),
-}
-
-fn handle_action(
-    action: KeyAction,
-    state: &mut AppState,
-    navigator: &mut TreeNavigator,
-    focused_path: &Option<PathBuf>,
-    entries: &[EntrySnapshot],
-    config: &Config,
-    text_preview: &mut Option<TextPreview>,
-) -> anyhow::Result<ActionResult> {
-    match action {
-        KeyAction::None => {}
-        KeyAction::Quit => {
-            state.should_quit = true;
-        }
-        KeyAction::Cancel => {
-            match &state.mode {
-                ViewMode::Browse => {
-                    if state.pick_mode {
-                        // Cancel in pick mode = exit with cancelled code
-                        return Ok(ActionResult::Quit(exit_code::CANCELLED));
-                    }
-                    state.should_quit = true;
-                }
-                _ => {
-                    state.mode = ViewMode::Browse;
-                    state.clear_message();
-                }
-            }
-        }
-        KeyAction::MoveUp => {
-            state.focus_index = state.focus_index.saturating_sub(1);
-        }
-        KeyAction::MoveDown => {
-            if state.focus_index < entries.len().saturating_sub(1) {
-                state.focus_index += 1;
-            }
-        }
-        KeyAction::MoveToTop => {
-            state.focus_index = 0;
-        }
-        KeyAction::MoveToBottom => {
-            state.focus_index = entries.len().saturating_sub(1);
-        }
-        KeyAction::Expand => {
-            if let Some(path) = focused_path {
-                navigator.expand(path)?;
-            }
-        }
-        KeyAction::Collapse => {
-            if let Some(path) = focused_path {
-                navigator.collapse(path);
-            }
-        }
-        KeyAction::ToggleExpand => {
-            if state.preview_visible {
-                // Close side preview panel
-                state.preview_visible = false;
-            } else if let Some(ref path) = focused_path {
-                if path.is_dir() {
-                    navigator.toggle_expand(path)?;
-                } else {
-                    // File: open fullscreen preview
-                    state.mode = ViewMode::Preview { scroll: 0 };
-                }
-            }
-        }
-        KeyAction::CollapseAll => {
-            // Collapse all except root
-            let entries_to_collapse: Vec<_> = entries
-                .iter()
-                .filter(|e| e.is_dir && e.depth > 0)
-                .map(|e| e.path.clone())
-                .collect();
-            for path in entries_to_collapse {
-                navigator.collapse(&path);
-            }
-        }
-        KeyAction::ExpandAll => {
-            // Expand all directories (limited depth to avoid huge trees)
-            let entries_to_expand: Vec<_> = entries
-                .iter()
-                .filter(|e| e.is_dir && e.depth < 5)
-                .map(|e| e.path.clone())
-                .collect();
-            for path in entries_to_expand {
-                navigator.expand(&path)?;
-            }
-        }
-        KeyAction::ToggleMark => {
-            if let Some(path) = focused_path {
-                if state.selected_paths.contains(path) {
-                    state.selected_paths.remove(path);
-                } else {
-                    state.selected_paths.insert(path.clone());
-                }
-            }
-        }
-        KeyAction::ClearMarks => {
-            state.selected_paths.clear();
-        }
-        KeyAction::Copy => {
-            let paths: Vec<PathBuf> = if state.selected_paths.is_empty() {
-                focused_path.clone().into_iter().collect()
-            } else {
-                state.selected_paths.iter().cloned().collect()
-            };
-            if !paths.is_empty() {
-                let mut clipboard = Clipboard::new();
-                let count = paths.len();
-                clipboard.copy(paths);
-                state.clipboard = Some(clipboard);
-                state.set_message(format!("Copied {} item(s)", count));
-            }
-        }
-        KeyAction::Cut => {
-            let paths: Vec<PathBuf> = if state.selected_paths.is_empty() {
-                focused_path.clone().into_iter().collect()
-            } else {
-                state.selected_paths.iter().cloned().collect()
-            };
-            if !paths.is_empty() {
-                let mut clipboard = Clipboard::new();
-                let count = paths.len();
-                clipboard.cut(paths);
-                state.clipboard = Some(clipboard);
-                state.set_message(format!("Cut {} item(s)", count));
-            }
-        }
-        KeyAction::Paste => {
-            if let Some(ref mut clipboard) = state.clipboard {
-                if let Some(content) = clipboard.take() {
-                    let dest = get_target_directory(focused_path.as_ref(), &state.root);
-
-                    match content {
-                        ClipboardContent::Copy(paths) => {
-                            for src in &paths {
-                                file_ops::copy_to(src, &dest)?;
-                            }
-                            state.set_message(format!("Pasted {} item(s)", paths.len()));
-                        }
-                        ClipboardContent::Cut(paths) => {
-                            for src in &paths {
-                                if let Some(name) = src.file_name() {
-                                    let new_path = dest.join(name);
-                                    std::fs::rename(src, new_path)?;
-                                }
-                            }
-                            state.set_message(format!("Moved {} item(s)", paths.len()));
-                        }
-                    }
-                    navigator.reload()?;
-                    state.refresh_git_status();
-                }
-            }
-        }
-        KeyAction::ConfirmDelete => {
-            let targets = create_delete_targets(state, focused_path.as_ref());
-            if !targets.is_empty() {
-                state.mode = ViewMode::Confirm {
-                    action: PendingAction::Delete { targets },
-                };
-            }
-        }
-        KeyAction::ExecuteDelete => {
-            if let ViewMode::Confirm {
-                action: PendingAction::Delete { targets },
-            } = &state.mode
-            {
-                for path in targets {
-                    file_ops::delete(path)?;
-                }
-                state.set_message(format!("Deleted {} item(s)", targets.len()));
-                state.selected_paths.clear();
-                state.mode = ViewMode::Browse;
-                navigator.reload()?;
-                state.refresh_git_status();
-            }
-        }
-        KeyAction::StartRename => {
-            if let Some(path) = focused_path {
-                let name = get_filename_str(Some(path));
-                state.mode = ViewMode::Input {
-                    purpose: InputPurpose::Rename {
-                        original: path.clone(),
-                    },
-                    buffer: name.clone(),
-                    cursor: name.len(),
-                };
-            }
-        }
-        KeyAction::StartNewFile => {
-            state.mode = ViewMode::Input {
-                purpose: InputPurpose::CreateFile,
-                buffer: String::new(),
-                cursor: 0,
-            };
-        }
-        KeyAction::StartNewDir => {
-            state.mode = ViewMode::Input {
-                purpose: InputPurpose::CreateDir,
-                buffer: String::new(),
-                cursor: 0,
-            };
-        }
-        KeyAction::StartSearch => {
-            state.mode = ViewMode::Search {
-                query: String::new(),
-            };
-        }
-        KeyAction::SearchNext => {
-            if let ViewMode::Search { query } = &state.mode {
-                if !query.is_empty() {
-                    let query_lower = query.to_lowercase();
-                    // Find next match starting from current position
-                    let start = (state.focus_index + 1) % entries.len();
-                    for i in 0..entries.len() {
-                        let idx = (start + i) % entries.len();
-                        if entries[idx].name.to_lowercase().contains(&query_lower) {
-                            state.focus_index = idx;
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-        KeyAction::Refresh => {
-            navigator.reload()?;
-            state.refresh_git_status();
-            state.set_message("Refreshed");
-        }
-        KeyAction::ToggleHidden => {
-            state.show_hidden = !state.show_hidden;
-            navigator.set_show_hidden(state.show_hidden)?;
-            state.set_message(if state.show_hidden {
-                "Showing hidden files"
-            } else {
-                "Hiding hidden files"
-            });
-        }
-        KeyAction::CopyPath => {
-            if let Some(path) = focused_path {
-                match arboard::Clipboard::new()
-                    .and_then(|mut cb| cb.set_text(path.display().to_string()))
-                {
-                    Ok(_) => state.set_message("Path copied to clipboard"),
-                    Err(_) => state.set_message("Failed to copy path to clipboard"),
-                }
-            }
-        }
-        KeyAction::CopyFilename => {
-            if let Some(path) = focused_path {
-                let name = get_filename_str(Some(path));
-                match arboard::Clipboard::new().and_then(|mut cb| cb.set_text(name)) {
-                    Ok(_) => state.set_message("Filename copied to clipboard"),
-                    Err(_) => state.set_message("Failed to copy filename to clipboard"),
-                }
-            }
-        }
-        KeyAction::OpenPreview => {
-            if matches!(state.mode, ViewMode::Preview { .. }) {
-                state.mode = ViewMode::Browse;
-            } else {
-                state.mode = ViewMode::Preview { scroll: 0 };
-            }
-        }
-        KeyAction::ToggleQuickPreview => {
-            state.preview_visible = !state.preview_visible;
-        }
-        KeyAction::ConfirmInput { value } => {
-            match &state.mode {
-                ViewMode::Input { purpose, .. } => {
-                    let parent = get_target_directory(focused_path.as_ref(), &state.root);
-                    match purpose {
-                        InputPurpose::CreateFile => {
-                            file_ops::create_file(&parent, &value)?;
-                            navigator.reload()?;
-                            state.refresh_git_status();
-                            state.set_message(format!("Created file: {}", value));
-                        }
-                        InputPurpose::CreateDir => {
-                            file_ops::create_dir(&parent, &value)?;
-                            navigator.reload()?;
-                            state.refresh_git_status();
-                            state.set_message(format!("Created directory: {}", value));
-                        }
-                        InputPurpose::Rename { original } => {
-                            file_ops::rename(original, &value)?;
-                            navigator.reload()?;
-                            state.refresh_git_status();
-                            state.set_message(format!("Renamed to: {}", value));
-                        }
-                    }
-                    state.mode = ViewMode::Browse;
-                }
-                ViewMode::Search { .. } => {
-                    // Keep search mode active, just update
-                    state.mode = ViewMode::Search { query: value };
-                }
-                _ => {}
-            }
-        }
-        KeyAction::PreviewScrollUp => {
-            if let Some(ref mut tp) = text_preview {
-                tp.scroll = tp.scroll.saturating_sub(1);
-            }
-            if let ViewMode::Preview { scroll } = &mut state.mode {
-                *scroll = scroll.saturating_sub(1);
-            }
-        }
-        KeyAction::PreviewScrollDown => {
-            if let Some(ref mut tp) = text_preview {
-                tp.scroll += 1;
-            }
-            if let ViewMode::Preview { scroll } = &mut state.mode {
-                *scroll += 1;
-            }
-        }
-        KeyAction::PreviewPageUp => {
-            if let Some(ref mut tp) = text_preview {
-                tp.scroll = tp.scroll.saturating_sub(20);
-            }
-            if let ViewMode::Preview { scroll } = &mut state.mode {
-                *scroll = scroll.saturating_sub(20);
-            }
-        }
-        KeyAction::PreviewPageDown => {
-            if let Some(ref mut tp) = text_preview {
-                tp.scroll += 20;
-            }
-            if let ViewMode::Preview { scroll } = &mut state.mode {
-                *scroll += 20;
-            }
-        }
-        KeyAction::PreviewToTop => {
-            if let Some(ref mut tp) = text_preview {
-                tp.scroll = 0;
-            }
-            if let ViewMode::Preview { scroll } = &mut state.mode {
-                *scroll = 0;
-            }
-        }
-        KeyAction::PreviewToBottom => {
-            if let Some(ref mut tp) = text_preview {
-                tp.scroll = tp.lines.len().saturating_sub(20);
-            }
-        }
-        KeyAction::PickSelect => {
-            if state.pick_mode {
-                let paths: Vec<PathBuf> = if state.selected_paths.is_empty() {
-                    focused_path.clone().into_iter().collect()
-                } else {
-                    state.selected_paths.iter().cloned().collect()
-                };
-
-                if !paths.is_empty() {
-                    // Execute callback if configured
-                    if let Some(ref callback) = config.callback {
-                        for path in &paths {
-                            let _ = callback.execute(path);
-                        }
-                    }
-
-                    // Output paths
-                    let result = PickResult::Selected(paths);
-                    return Ok(ActionResult::Quit(result.output(config.output_format)?));
-                }
-            }
-        }
-        KeyAction::ShowHelp => {
-            state.set_message("j/k:move l/h:expand/collapse Space:mark y/d/p:copy/cut/paste D:delete a/A:new r:rename /:search ?:help");
-        }
-    }
-
-    Ok(ActionResult::Continue)
 }
